feat: move FilterByDate section to AllTransactions Screen package
